frontend: move account guide to its own component
Moved isBTCScriptType to utils and the account guide to its own
component so that account.tsx is slightly simpler and easier to
refactor.
